feat: deploy mumbai GSN contract

+++ contracts/DankBankMarketData.sol
@@ -5,32 +5,17 @@ contract DankBankMarketData {
     mapping(address => uint256) public virtualEthPoolSupply;
     mapping(address => uint256) public ethPoolSupply;
 
-    event LiquidityAdded(
-        address indexed funder,
-        address token,
-        uint256 amountAdded,
-        uint256 sharesMinted
-    );
+    event LiquidityAdded(address indexed funder, address token, uint256 amountAdded, uint256 sharesMinted);
 
     event LiquidityRemoved(
         address indexed funder,
         address token,
         uint256 tokensRemoved,
         uint256 ethRemoved,
-        uint sharesBurnt
+        uint256 sharesBurnt
     );
 
-    event DankBankBuy(
-        address indexed buyer,
-        address token,
-        uint256 investmentAmount,
-        uint256 tokensBought
-    );
+    event DankBankBuy(address indexed buyer, address token, uint256 investmentAmount, uint256 tokensBought);
 
-    event DankBankSell(
-        address indexed seller,
-        address token,
-        uint256 returnAmount,
-        uint256 tokensSold
-    );
+    event DankBankSell(address indexed seller, address token, uint256 returnAmount, uint256 tokensSold);
 }
